+++ @xen-orchestra/lite/src/App/TabConsole.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { withState } from 'reaclette'
 
+import Button from '../components/Button'
 import Console from '../components/Console'
-import RangeInput from '../components/RangeInput'
 
 interface ParentState {}
 
 interface State {
-  consoleScale: number
+  ctrlAltDel: () => void
 }
 
 interface Props {
@@ -17,7 +17,7 @@ interface Props {
 interface ParentEffects {}
 
 interface Effects {
-  scaleConsole: React.ChangeEventHandler<HTMLInputElement>
+  setCtrlAltDel: (fn: () => void) => void
 }
 
 interface Computed {}
@@ -25,26 +25,18 @@ interface Computed {}
 const TabConsole = withState<State, Props, Effects, Computed, ParentState, ParentEffects>(
   {
     initialState: () => ({
-      // Value in percent
-      consoleScale: 100,
+      ctrlAltDel: () => {},
     }),
     effects: {
-      scaleConsole: function (e) {
-        this.state.consoleScale = +e.currentTarget.value
-
-        // With "scaleViewport", the canvas occupies all available space of its
-        // container. But when the size of the container is changed, the canvas
-        // size isn't updated
-        // Issue https://github.com/novnc/noVNC/issues/1364
-        // PR https://github.com/novnc/noVNC/pull/1365
-        window.dispatchEvent(new UIEvent('resize'))
+      setCtrlAltDel: function (fn) {
+        this.state.ctrlAltDel = fn
       },
     },
   },
-  ({ state, effects, vmId }) => (
+  ({ effects, state, vmId }) => (
     <div style={{ height: '100vh' }}>
-      <RangeInput max={100} min={1} onChange={effects.scaleConsole} step={1} value={state.consoleScale} />
-      <Console vmId={vmId} scale={state.consoleScale} />
+      <Button label='CTRL+ALT+DEL' onClick={state.ctrlAltDel} />
+      <Console vmId={vmId} setCtrlAltDel={effects.setCtrlAltDel} />
     </div>
   )
 )
